+++ backend/canisters/identity/api/src/queries/auth_principals.rs
@@ -19,7 +19,6 @@ pub struct UserPrincipal {
     pub principal: Principal,
     pub originating_canister: CanisterId,
     pub is_ii_principal: bool,
-    pub is_current_identity: bool,
     pub webauthn_key: Option<WebAuthnKey>,
     pub last_used: TimestampMillis,
 }